pyproject.toml
[project]
name = "my-simple-repo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]>=0.135.1",
    "pillow>=12.1.1",
    "requests>=2.32.5",
]
